KSENROLL-14743 Create IdentifyChangesMojo
Given two ref names (assumed to be local git branches) we do a tree walk
looking for object id differences.

We know which files changed.

We also report on the modules (top level directories, . for the top) that
contain changes.

We also report on the modules (top level directories) that contain sql changed
files.

Right now we just report that the file changed not whether it is a create or
modify.