KSENROLL-14743 Create base class for authenticating to Github
In Jenkins we can use the EnvInject or Credentials Binding Plugins to setup
environment variables that contain the github username and password.

These credentials can be used for github api operations or Git operations.

To start with I've created a way to resolve the github api credentials using
either the ~/.github file or from the environment.