KSENROLL-14743 Generate one property file per pull request
In order to make the generated report more easily consumed by the Jenkins
Parameterized Trigger Plugin I've changed how the report of open pull requests
is created.

Now a file is created for each open pull request.  It conforms to a java
properties file where there are key=value pairs on each line.

The naming of the parameters is aligned with what the jenkins job is expecting.

Also update the unit test to run without the presense of the ~/.github file.