KSENROLL-14743 Add support for environment passthrough in api change detection
Additional environment variables like the pull request commit id may be
available in the environment.

We don't have to use them here but they are valuable for the downstream jobs
created from the generated files.

So add a mechanism where the plugin can be configured with the names of
environment variables which if they exist should be included in the generated
files.