KSENROLL-14743 Several Mojo Adjustments.
Create a new PushGitReferenceMojo that will configure on disk C git credentials
that will allow pushing up certain refspec's from the CI repository.

Auto calculate the depth of the shallow clone from within the
FetchOpenPullRequestsMojo.

Given a Development, D branch with topic branches T_a and T_b:

      T_b
     /
D -> D -> D
\
 T_a -> T_a

The depth of T_a is 2, the depth of T_b is 1 and the depth of the development
branch is 3.  We choose the max of these values which is 3.

Added CheckCrossProjectBranchContentThroughApiMojo which will look for certain
strings in commit messages.  This will be used to find out which head revision
the impex changes applied to.

Now that we have the full graph we can do an analysis of each element of the
changes to see when the impex change occured.